linear search iteration (Find index)
Wrote a functiom to perform a linear search on an unsorted lisr of numbers. If target is found, return its index otherwise return -1. Wrote function using correct syntax. Made afew mistakes but got there in the end. Indexs start at 0 so since 40 is the 4th element but at 3rd index, the result is index 3.